model/requirements.txt
datasets
numpy
transformers
torchvision
supervision==0.10.0
pycocotools
pytorch_lightning
opencv-python-headless
opencv-python
yt-dlp
Pillow
torch